// apps/cms/src/shared/api/generated/notifications/schemas/listNotificationTemplateVariablesMemberType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Notifications subset
 * Filtered for CMS notifications / kakao alimtalk Orval codegen.
 * OpenAPI spec version: v9
 */

export type ListNotificationTemplateVariablesMemberType = typeof ListNotificationTemplateVariablesMemberType[keyof typeof ListNotificationTemplateVariablesMemberType];


export const ListNotificationTemplateVariablesMemberType = {
  GENERAL: 'GENERAL',
  SCHOOL_TEACHER: 'SCHOOL_TEACHER',
  INSTRUCTOR: 'INSTRUCTOR',
  TEACHER_AND_INSTRUCTOR: 'TEACHER_AND_INSTRUCTOR',
  ADMIN: 'ADMIN',
} as const;