// apps/cms/src/shared/api/generated/notifications/schemas/listNotificationTemplateVariablesParticipantType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Notifications subset
 * Filtered for CMS notifications / kakao alimtalk Orval codegen.
 * OpenAPI spec version: v9
 */

export type ListNotificationTemplateVariablesParticipantType = typeof ListNotificationTemplateVariablesParticipantType[keyof typeof ListNotificationTemplateVariablesParticipantType];


export const ListNotificationTemplateVariablesParticipantType = {
  PARTICIPANT: 'PARTICIPANT',
  INSTRUCTOR: 'INSTRUCTOR',
  VOLUNTEER: 'VOLUNTEER',
} as const;